Moved Modal to the PaymentMethod component.
